Create README1.md
Medical Reminder App

This is a mobile app for elderly people above 65 years with a use case of medical reminder built using HTML, CSS, and JavaScript. The app allows users to add medications and reminders, and displays a list of added medications and reminders.

Getting Started

To get started with the app, simply open the "index.html" file in any web browser. The app will load automatically and you can start using it immediately.

Usage

To add a medication, fill out the medication form with the medication name and dosage and click the "Add Medication" button. The medication will be added to the medication list.

To add a reminder, fill out the reminder form with the reminder name and time and click the "Add Reminder" button. The reminder will be added to the reminder list.

To view the list of added medications and reminders, simply scroll down to the corresponding section of the app.

License

This app is licensed under the MIT License. Feel free to use, modify, and distribute the app as you see fit. However, please note that the app is provided "as is" and the author assumes no liability for any damages or losses arising from its use.